// src/components/forms/product/CreateProduct/types.ts
export interface IProductInfo {
  name?: string;
  brand?: string;
  size?: string;
  sizeCountry?: string;
  description?: string;
  additionalMaterials?: string[];
}

export interface ISpecifications {
  weight?: string;
  usage?: string;
  top?: string;
  sole?: string;
  features?: string;
  lacingMode?: string;
  soleType?: string;
  sockType?: string;
  mainColor?: string;
  height?: string;
  season?: string;
  article?: string;
  officialPrice?: string;
  releaseDate?: string;
  set?: string;
  materials?: string;
  style?: string;
  category?: string;
}